# docs/v1/openapi.yaml
...
openapi: 3.0.0
info:
  title: Verifiable Credential Service (VCS)
  version: 0.2.0
  description: TrustBloc VCS support credential issuance and verification using W3C VC-API along with OpenID4VC protocol for issuer/verifier interaction with Wallet.
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.txt'
  contact:
    name: TrustBloc team.
    email: hello@trustbloc.dev
    url: 'https://trustbloc.dev'
servers:
  - url: 'http://localhost:8070'
...
paths:
...
  /oidc/par:
    post:
      summary: OIDC Pushed Authorization Request
      tags:
        - oidc4vc
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushedAuthorizationResponse'
        '405':
          description: Method Not Allowed
        '413':
          description: Request Entity Too Large
        '429':
          description: Too Many Requests
      operationId: oidc-pushed-authorization-request
      description: Client sends OAuth authorization request directly to authorization server and gets request URI in response that can be used as reference to the data in subsequent request to authorization endpoint.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                response_type:
                  type: string
                  description: Value MUST be set to "code".
                client_id:
                  type: string
                  description: The client identifier.
                code_challenge:
                  type: string
                  description: 'A challenge derived from the code verifier that is sent in the authorization request, to be verified against later.'
                code_challenge_method:
                  type: string
                  description: A method that was used to derive code challenge.
                redirect_uri:
                  type: string
                  description: The authorization server redirects the user-agent to the client's redirection endpoint previously established with the authorization server during the client registration process or when making the authorization request.
                scope:
                  type: string
                  description: The scope of the access request.
                state:
                  type: string
                  description: An opaque value used by the client to maintain state between the request and callback. The authorization server includes this value when redirecting the user-agent back to the client. The parameter SHOULD be used for preventing cross-site request forgery.
                authorization_details:
                  type: string
                  description: The authorization_details conveys the details about the credentials the wallet wants to obtain. Multiple authorization_details can be used with type openid_credential to request authorization in case of multiple credentials.
                wallet_issuer:
                  type: string
                  description: Wallet's OpenID Connect Issuer URL. The Issuer will use the discovery process to determine the wallet's capabilities and endpoints. RECOMMENDED in Dynamic Credential Request.
                user_hint:
                  type: string
                  description: An opaque user hint the wallet MAY use in subsequent callbacks to optimize the user's experience. RECOMMENDED in Dynamic Credential Request.
                op_state:
                  type: string
                  description: 'String value identifying a certain processing context at the credential issuer. A value for this parameter is typically passed in an issuance initiation request from the issuer to the wallet. This request parameter is used to pass the  op_state value back to the credential issuer. The issuer must take into account that op_state is not guaranteed to originate from this issuer, could be an attack.'
              required:
                - response_type
                - client_id
                - code_challenge
        description: ''
  /oidc/pre-authorize:
    post:
      summary: Request token for Pre-Authorized flow.
      tags:
        - oidc4vc
      description: Used for Pre-Authorized flow to request access token with pre-auth code and optional pin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreAuthorizeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
  /oidc/authorize:
    get:
      summary: OIDC Authorization Request
      responses:
        '302':
          description: Found
          headers: { }
          content:
            application/x-www-form-urlencoded:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    description: REQUIRED. The authorization code generated by the authorization server. The authorization code MUST expire shortly after it is issued to mitigate the risk of leaks.
                  state:
                    type: string
                    description: REQUIRED if the "state" parameter was present in the client authorization request. The exact value received from the client.
                required:
                  - code
      operationId: oidc-authorize
      description: 'OAuth 2.0 Authorization Request, which requests to grant access to the Credential endpoint.'
      parameters:
        - schema:
            type: string
          in: query
          name: response_type
          description: Value MUST be set to "code".
          required: true
        - schema:
            type: string
          in: query
          name: client_id
          description: The client identifier.
          required: true
        - schema:
            type: string
          in: query
          name: code_challenge
          description: 'A challenge derived from the code verifier that is sent in the authorization request, to be verified against later.'
          required: true
        - schema:
            type: string
          in: query
          name: code_challenge_method
          description: A method that was used to derive code challenge.
        - schema:
            type: string
          in: query
          name: redirect_uri
          description: The authorization server redirects the user-agent to the client's redirection endpoint previously established with the authorization server during the client registration process or when making the authorization request.
        - schema:
            type: string
          in: query
          name: scope
          description: The scope of the access request.
        - schema:
            type: string
          in: query
          name: state
          description: An opaque value used by the client to maintain state between the request and callback. The authorization server includes this value when redirecting the user-agent back to the client. The parameter SHOULD be used for preventing cross-site request forgery.
        - schema:
            type: string
          in: query
          name: authorization_details
          description: The authorization_details conveys the details about the credentials the wallet wants to obtain. Multiple authorization_details can be used with type openid_credential to request authorization in case of multiple credentials.
        - schema:
            type: string
          in: query
          name: wallet_issuer
          description: Wallet's OpenID Connect Issuer URL. The Issuer will use the discovery process to determine the wallet's capabilities and endpoints. RECOMMENDED in Dynamic Credential Request.
        - schema:
            type: string
          in: query
          name: user_hint
          description: An opaque user hint the wallet MAY use in subsequent callbacks to optimize the user's experience. RECOMMENDED in Dynamic Credential Request.
        - schema:
            type: string
          in: query
          name: op_state
          description: 'String value identifying a certain processing context at the credential issuer. A value for this parameter is typically passed in an issuance initiation request from the issuer to the wallet. This request parameter is used to pass the  op_state value back to the credential issuer. The issuer must take into account that op_state is not guaranteed to originate from this issuer, could be an attack.'
          required: true
      tags:
        - oidc4vc
    parameters: [ ]
  /oidc/token:
    post:
      summary: OIDC Token Request
      tags:
        - oidc4vc
      operationId: oidc-token
      description: Issues access token and optionally a refresh token for the exchange of authorization code that client has obtained after successful authorization response.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '400':
          description: Bad Request
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                  description: Value MUST be set to "authorization_code".
                code:
                  type: string
                  description: The authorization code received from the authorization server.
                code_verifier:
                  type: string
                  description: A cryptographically random string that is used to correlate the authorization request to the token request.
                redirect_uri:
                  type: string
                  description: 'REQUIRED, if the "redirect_uri" parameter was included in the authorization request, and their values MUST be identical.'
                client_id:
                  type: string
                  description: 'REQUIRED, if the client is not authenticating with the authorization server.'
              required:
                - grant_type
                - code
                - code_verifier
        description: ''
  /oidc/redirect:
    get:
      summary: OIDC Redirect
      tags:
        - oidc4vc
      operationId: oidc-redirect
      description: OIDC redirect for handling response from issuer's OIDC provider and continue our OIDC authorize flow.
      responses:
        '303':
          description: Redirect
      parameters:
        - schema:
            type: string
          name: code
          in: query
          required: true
          description: auth code for issuer provider
        - schema:
            type: string
          name: state
          in: query
          required: true
          description: state
components:
  schemas:
...
    PreAuthorizeRequest:
      title: PreAuthorizeRequest
      x-tags:
        - issuer
      type: object
      description: Model exchanging pre-authorized code and pin to token.
      properties:
        pre_auth_code:
          type: string
        pin:
          type: string
      required:
        - pre_auth_code
...
    PushedAuthorizationResponse:
      title: PushedAuthorizationResponse
      x-tags:
        - oidc4vc
      type: object
      description: Model for Pushed Authorization Response.
      properties:
        request_uri:
          type: string
          description: The request URI corresponding to the authorization request posted. This URI is a single-use reference to the respective request data in the subsequent authorization request.
        expires_in:
          type: integer
          description: 'A JSON number that represents the lifetime of the request URI in seconds as a positive integer. The request URI lifetime is at the discretion of the authorization server but will typically be relatively short (e.g., between 5 and 600 seconds).'
      required:
        - request_uri
        - expires_in
    AccessTokenResponse:
      title: AccessTokenResponse
      x-tags:
        - oidc4vc
      type: object
      description: Model for Access Token Response.
      properties:
        access_token:
          type: string
          description: The access token issued by the authorization server.
        token_type:
          type: string
          description: The type of the token issued.
        expires_in:
          type: integer
          description: The lifetime in seconds of the access token.
        refresh_token:
          type: string
          description: 'The refresh token, which can be used to obtain new access tokens.'
        scope:
          type: string
          description: 'OPTIONAL, if identical to the scope requested by the client; otherwise, REQUIRED.'
        c_nonce:
          type: string
          description: String containing a nonce to be used to create a proof of possession of key material when requesting a credential.
        c_nonce_expires_in:
          description: Integer denoting the lifetime in seconds of the c_nonce.
          type: integer
      required:
        - access_token
        - token_type
...
  securitySchemes: { }